% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphcorrelation.R
\name{graphcorrelation}
\alias{graphcorrelation}
\title{Calculate epigenoproteomic correlations}
\usage{
graphcorrelation(
  Peak_annotation,
  PeakCount,
  TranscriptCount,
  ProteinCount,
  orderSamples,
  graphSamples
)
}
\arguments{
\item{Peak_annotation}{}

\item{PeakCount}{}

\item{TranscriptCount}{}

\item{ProteinCount}{}

\item{orderSamples}{}

\item{graphSamples}{}
}
\value{

}
\description{
Calculate epigenoproteomic correlations
}
\examples{
graphcorrelation(data/Peak_annotation.rda,data/PeakCount_NonPromoter.rda,data/TranscriptCount.rda,data/ProteinCount.rda,data/orderSamples.rda,c("1","2"))
}
