% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictprotein.R
\name{predictprotein}
\alias{predictprotein}
\title{}
\usage{
predictprotein(
  Peak_annotation,
  PeakCount,
  TranscriptCount,
  ProteinCount,
  orderSamples,
  testSamples
)
}
\arguments{
\item{Peak_annotation}{}

\item{PeakCount}{}

\item{TranscriptCount}{}

\item{ProteinCount}{}

\item{orderSamples}{}

\item{graphSamples}{}
}
\value{

}
\description{

}
\examples{
predictprotein(data/Peak_annotation.rda,data/PeakCount_NonPromoter.rda,data/TranscriptCount.rda,data/ProteinCount.rda,data/orderSamples.rda,c("1","2"))
}
