% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{randomize}
\alias{randomize}
\title{}
\usage{
randomize(orderSamples)
}
\arguments{
\item{orderSamples}{}
}
\value{

}
\description{

}
\examples{
randomize(c("1","2","3"))
randomize(data/orderSamples.rda)
}
